- separate workers in another module - some modifications
